[amlogic-display][hpd] Use async loops to handle hotplug IRQs.
Previously, display hotplug IRQs were handled directly by the
zx_interrupt_wait() syscall, on a dedicated IRQ handler thread.
This made it difficult to pause / resume IRQ handling.

This change replaces the interrupt waits with the async::Irq
handler. It binds the interrupt to an async loop (backed by
Zircon ports); interrupts are delivered as packets and handled
by the async IRQ handler running on the loop.
This change doesn't change the HotPlugDetection interface.

Besides, the driver framework v2 prefers drivers to dispatch
asynchronous work on the driver dispatcher over spawning their own
threads. The DFv2 driver dispatcher supports binding async::Irq
and IrqMethod handlers, so this change also makes it easier to migrate
the driver to DFv2 in the future.

Some special configuration of the async loop is needed to support
IRQ binding on the loop. We define this in the hot-plug-detection.cc
as "kIrqHandlerAsyncLoopConfig".

Test: No-op expected.
Test: amlogic-display-test

Test: "display-tool squares" on VIM3 with AML028-30MB-A1 with the
      following configurations:

      Configuration 1. display is plugged while VIM3 powers on.
      Tests passed repeatedly on VIM3.

      Configuration 2. Plug the display, run "display-tool squaures",
      then unplug the display. Repeat this for 10 times. All the
      tests passed.

Change-Id: Ideab6d2b0e2b7ec725aa42841ed28e23ac42b525
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/960154
Commit-Queue: Yilong Li <liyl@google.com>
Reviewed-by: Victor Costan <costan@google.com>