[component_debug] handle component destruction in get_all_route_segments
The `get_all_route_segments` function uses RealmQuery to produce a list
of all components, and then asks for the resolved information for each
component in the list. It's possible for a component to be destroyed
between these two steps, and thus the function must gracefully handle a
component not existing when it asks for its resolved information.

Multiply: host_x64/verify_ffx_profile_heapdump
Fixed: 316443204
Change-Id: I4003f5286afd0a21e4cd121f066a5345cef09365
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/968572
Reviewed-by: Fabio D'Urso <fdurso@google.com>
Commit-Queue: Claire Gonyeo <cgonyeo@google.com>