[compat] Fix flake with astro OTA test.
The OTA test on smart_display_eng.astro-release-e2e-isolated-omaha
started flaking after delaying calling release until after
dispatcher shutdown (fxr/961879).

fxr/965633 was an attempted fix to destruct as much of the device
as possible.

This CL adds the call to child_pre_release in the case where
we are delaying calling the release, and also remove the
calling of remove_completer earlier - this instead is called
in the device destructor.

Bug: 317784515

Change-Id: I8ea89762bc96469c70d05e43b7f29067017f5c46
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/968077
Commit-Queue: Jocelyn Dang <jocelyndang@google.com>
Reviewed-by: Nick Eaton <nickeaton@google.com>