[starnix] Optimize shared futex operations
In Linux futex operations can be shared (the default) or private. The
difference is that when a shared operation blocks it can be unblocked by
shared operations on other addresses that map to the same underlying
memory. This optimizes shared operations by loading the current value
directly from the thread's existing memory mapping when possible and
only computing a key for shared operations by looking up the backing VMO
when the operation has committed to blocking. This also updates the code
to always check for alignment before attempting a load rather than as a
side effect of computing a key (since we don't need to compute a key
unless we are blocking) and splits the external loading logic out.

Change-Id: I35455b55b04ad0218d927a47460b4e42a9ace1b1
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/967420
Fuchsia-Auto-Submit: James Robinson <jamesr@google.com>
Reviewed-by: Ghanan Gowripalan <ghanan@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>