[starnix] Use BStr(ing) for FsStr(ing).
These types are essentially &[u8] and Vec<u8> with
Debug/Display/Eq/etc implementations which assume the contents
are UTF-8 by convention and which gracefully handle invalid UTF-8
when printing or converting to Rust strings.

This should make it easier to debug and report errors in the future.

This change brings some extra line noise for conversions, especially
when using literals to interact with APIs that accept names and when
matching on FsStr(ing)s. I explored using `impl AsRef<[u8]>` for a
number of APIs to allow multiple types to be passed, but it could
increase compile times and many of the places we benefit from the
new trait implementations are within methods that would need to
be generic for convenience.

Bug: b/297369529
Change-Id: I58715ee3437f6ba80b999dd0d9a3520d63308ce6
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/969452
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>
Commit-Queue: Adam Perry <adamperry@google.com>
Reviewed-by: Vickie Cheng <vickiecheng@google.com>
Fuchsia-Auto-Submit: Adam Perry <adamperry@google.com>