[fidlc] Check that FIDL source files are valid UTF-8
This is a simple, check based on the logic in //src/lib/fxl. The error
reporting is simple - it'll just print the error "Not valid UTF-8". No
attempts to report byte or line offsets or guess that really you're
passing in a latin-9 file by mistake.

In the process, we now print the errno string for system errors involved
in opening source files (eg: "No such file or directory" for file not
found) and "Not a regular file" if you pass in a path that points to
something other than a file (eg: directory, socket, device, etc).

There are no tests because this code is at the very front-end of the
command-line tools which we haven't built a way to test. But it's simple
and calls into code that is well tested.

Fixes: b/42064369
Change-Id: If4ba7ed4f9d1ff6adfc41680c065ca20680deeca
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/969780
Reviewed-by: Clayton Wilkinson <wilkinsonclay@google.com>
Commit-Queue: Ian McKellar <ianloic@google.com>