[netconform] Add cobalt to assembly
Even though cobalt is not needed for network conformance testing,
the incessant logspam due to its absence makes it harder to debug
real issues, so let's add it to the assembly to quiet things.

Achieve this by raising the assembly feature_set_level to the default
"minimal" rather than "utility" so that it's automatically included in
the assembly.

Change-Id: I27c7ceb3b8468c8e7107266c6f07a087fe485f77
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/968675
Reviewed-by: Bruno Dal Bo <brunodalbo@google.com>
Commit-Queue: Evan Wildenhain <wildenhain@google.com>