[starnix][power] suspend cpu end2end test
Create an end2end test like squareload to make sure the CPUs suspend
and resume correctly when /sys/power/state is written.

Note: The test is failing to parse the trace, possibly due to unmatched
duration events. Since we have not published it to any CI/CQ trybot,
it is safe to submit it to check on the power graph in a communal try
builder. The follow-up fix on performance trace will be tracked in
b/317380861.

Bug: b/317380861
Change-Id: Ia769e4f31641a2763119e558b8072a7772a9a5da
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/968762
Commit-Queue: Guocheng Wei <guochengwei@google.com>
Reviewed-by: Vinicius Felizardo <felizardo@google.com>
Reviewed-by: Kevin Cho <chok@google.com>