[ffx][plugins] Adding explicit anyhow dependency
This adds anyhow as a dependency to plugins that directly
reference the anyhow crate. Previously, the ffx_plugin GN template
silently added a dependency on anyhow.

This CL is in preparation for changing the implementation of the
ffx_plugin to not use anyhow, removing the silent addition of it as a
dependency.

Bug: b/287782180
Change-Id: Id7772033d837857f3939c1e4bac66a32ed45d9f0
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/969785
Reviewed-by: Dwayne Slater <dwayneslater@google.com>
Commit-Queue: Clayton Wilkinson <wilkinsonclay@google.com>