[ffx][fho] Improve error message for device timeout
This improves the error message when there is a timeout connecting
to a device to establish the remotecontrol proxy connection.

Previously, the error was being wrapped in an anyhow::Error which
caused it to be printed as an unknown error with the "BUG:..." prefix.
This removes the wrapper so the error can be processed as expected.

Fixes: b/318882239
Change-Id: Ie7a1aefd2db88b2a4355bccd867c423b6728783f
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/970146
Commit-Queue: Clayton Wilkinson <wilkinsonclay@google.com>
Reviewed-by: Steven Grady <slgrady@google.com>