[graphics] Explicitly depend on syslog
Previously, this dep was added implicitly through
fxl. This dependency needs to be removed in order
for fxl to be properly usable in drivers or other
contexts where syslog is not available.
This CL is part of an LSC to explicitly depend on syslog
where the dependency was previously brought in through
fxl.

Bug: b/300157652
Change-Id: Ic3d0a73f73b13c06bdf237f137e0d0713e7662ab
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/971354
Reviewed-by: John Bauman <jbauman@google.com>
Commit-Queue: Brian Bosak <bbosak@google.com>