[fidl] Explicitly depend on syslog
Previously, this dep was added implicitly through
fxl. This dependency needs to be removed in order
for fxl to be properly usable in drivers or other
contexts where syslog is not available.
This CL is part of an LSC to explicitly depend on syslog
where the dependency was previously brought in through
fxl.

Bug: b/300157652
Change-Id: I0a078b1f6b3a40ddcc556f3e9d2f74adfc38dd1f
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/971352
Reviewed-by: Mitchell Kember <mkember@google.com>
Commit-Queue: Brian Bosak <bbosak@google.com>