[fdf][testing] Shutdown dispatchers + stop after
In the real driver host, driver dispatchers that are created by the
driver are automatically shut down. This adds the same behavior in the
testing framework.

Another difference with the real environment was that the Stop hook is
called after the driver dispatcher has shutdown. This change makes it so
that the user can do the same thing if they choose by manually calling
the shutdown dispatchers function so they can call Stop manually after
it.

Also fixes an issue in the completer where it would call the completer
function, which could race and end up not allowing the release to
happen, which would cause the destructor assert to fail.

Change-Id: I1c58614e43455534e34be2ebfcfa3fa68bca1efc
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/970139
Reviewed-by: Zhiyi Chen <zhiyichen@google.com>
Reviewed-by: Karthik Krishnaswami <karthikrish@google.com>
Commit-Queue: Novin Changizi <novinc@google.com>
API-Review: Jocelyn Dang <jocelyndang@google.com>
Reviewed-by: Jocelyn Dang <jocelyndang@google.com>