[ui] Explicitly depend on syslog
Previously, this dep was added implicitly through
fxl. This dependency needs to be removed in order
for fxl to be properly usable in drivers or other
contexts where syslog is not available.
This CL is part of an LSC to explicitly depend on syslog
where the dependency was previously brought in through
fxl.

Bug: b/300157652
Change-Id: I03eb618cb1c9c84abd7f9d44d1bb7fc5bf0bcee2
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/971353
Reviewed-by: Jianpeng Chao <chaopeng@google.com>
Reviewed-by: Mukesh Agrawal <quiche@google.com>
Commit-Queue: Brian Bosak <bbosak@google.com>