[gesture_manager] Explicitly depend on syslog
Previously, this dep was added implicitly through
fxl. This dependency needs to be removed in order
for fxl to be properly usable in drivers or other
contexts where syslog is not available.
This CL is part of an LSC to explicitly depend on syslog
where the dependency was previously brought in through
fxl.

Bug: b/300157652
Change-Id: I80ff291c28b86356c02ecf68f98eea1174b7f352
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/971444
Commit-Queue: Brian Bosak <bbosak@google.com>
Reviewed-by: Mukesh Agrawal <quiche@google.com>