[components] Remove program_receivers_dict
This CL rips out the last Receiver<WeakComponentInstance> in the program
output routing path. Instead of some final entity watching receivers and
calling open, the full routing chain is performed via a series of Router
end-to-end, with the final Router returning an Open that is downscoped
from the outgoing directory of the program.

There are a lot of things going on when connecting to a component. As a
result this CL makes the following terminology distinction to clarify:

- program outgoing dict: a dict of routers that will request
  capabilities from the program, assuming it is already started

- program output dict: wraps the former with logic to check policy, then
  start the component

I put the various builder functions in component.rs in
ResolvedInstanceState to minimize code delta. Maybe some of them could
be instance methods on ComponentInstance, or moved to
sandbox_construction. I'm happy to apply those edits following the CL.

Run-All-Tests: True
Bug: b/310741884
Fixed: b/318568196

Change-Id: I2ecfd75adc65df1ea3453e33885574f1996358c8
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/966583
Fuchsia-Auto-Submit: Yifei Teng <yifeit@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>
Reviewed-by: Claire Gonyeo <cgonyeo@google.com>