[netstack3] Introduce DeviceSocketApi struct
Replace the `DeviceSocketApi` trait with a struct with the same name
which is exposed from `CoreApi`.

Eliminate the newtype for `SocketId` and make it a pure alias to match
the new pattern.

Move `ContextPair::core_api` to be agnostic on how SyncCtx is stored
through the `StackStateProvider` trait. That allows the loom tests to
share `SyncCtx` between threads and still use the same API as
testutil::Ctx.

Bug: b/42083910
Change-Id: I75e98209aca958cd4db907ab3ed5633017fc2eb3
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/971343
Reviewed-by: Jeff Martin <martinjeffrey@google.com>
Fuchsia-Auto-Submit: Bruno Dal Bo <brunodalbo@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>