[ui] Explicitly depend on syslog
Previously, this dep was added implicitly through
fxl. This dependency needs to be removed in order
for fxl to be properly usable in drivers or other
contexts where syslog is not available.
This CL is part of an LSC to explicitly depend on syslog
where the dependency was previously brought in through
fxl.

Bug: b/300157652
Change-Id: Ieaab562b49b54c2de3e04f8ca12b6c124166a353
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/971952
Reviewed-by: Jianpeng Chao <chaopeng@google.com>
Commit-Queue: Brian Bosak <bbosak@google.com>
Reviewed-by: Mukesh Agrawal <quiche@google.com>