[ffx][goldens] Improving ffx command golden build error
When the golden file test fails for the ffx command line,
the standard error from the build refers to the label of the
specific golden file test. In the case of ffx command line,
this only part of the overall test, so depending on it directly
causes the goldens to be incorrect.

This fixes the situation by restricting the visibility of the
sub-test labels so the correct fix is to depend on
//src/developer/ffx/tests/cli-goldens:tests.

This also adds a warning for the case where ffx_build_dual_mode_plugins_as_subtools
is set, which disables the test.

Fixes: b/319293919
Change-Id: I44975597d500c01203534ca9f3cb1ff3b9532cff
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/971434
Commit-Queue: Clayton Wilkinson <wilkinsonclay@google.com>
Reviewed-by: Colin Nelson <colnnelson@google.com>