[fidlcat] Explicitly depend on syslog
Previously, this dep was added implicitly through
fxl. This dependency needs to be removed in order
for fxl to be properly usable in drivers or other
contexts where syslog is not available.
This CL is part of an LSC to explicitly depend on syslog
where the dependency was previously brought in through
fxl.

Bug: b/300157652
Change-Id: I1572ecf3903e2f11c5db0d0a262f050af58f9440
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/972122
Commit-Queue: Brian Bosak <bbosak@google.com>
Reviewed-by: Jeremy Manson <jeremymanson@google.com>