[component_manager] Exposed dir sets dirent type
We also have to update the implementation of
RemoteControl/OpenCapability since it uses readdir_recursive. If we did
nothing, it would break the existence check for directory capabilities
since it would recurse into the directory. Instead, open the parent
directory of `capability_name` and read the entries in that without
recursing.

Fixed: 42076807
Change-Id: If5f45f150c85c436c1ecaf806007e9ab67dea349
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/970836
Reviewed-by: Andrew Davies <awdavies@google.com>
Commit-Queue: Gary Bressler <geb@google.com>
Reviewed-by: Yegor Pomortsev <ypomortsev@google.com>