[virtualization] Explicitly depend on syslog
Previously, this dep was added implicitly through
fxl. This dependency needs to be removed in order
for fxl to be properly usable in drivers or other
contexts where syslog is not available.
This CL is part of an LSC to explicitly depend on syslog
where the dependency was previously brought in through
fxl.

Bug: b/300157652
Change-Id: Ib845951bfeb091548383b1e4b88a5f126e6c36cd
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/972121
Reviewed-by: Will Drewry <drewry@google.com>
Reviewed-by: Tim Kilbourn <tkilbourn@google.com>
Commit-Queue: Brian Bosak <bbosak@google.com>