[fxfs] Fix bug where pages may not be supplied.
It's possible that VMO size and 'byte_size' are different.
This fixes a bug where, if byte_size is smaller, the page handler
may fail to supply required pages.

It also ensures that read ahead will not supply zero pages beyond the end
of the VMO (uses max of byte_size and requested range) as this
will cause supply_pages to fail a range check and lead to a similar
result.

Bug: b/318773519
Change-Id: I87dd37a7065a6e454668b7e72ad1f7a98992b1e4
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/971633
Commit-Queue: Aaron Drew <ripper@google.com>
Reviewed-by: Chris Suter <csuter@google.com>