[build][component] Add the syslog shard in fuchsia_unittest_component
Folks noticed  logs missing in a test using fuchsia_unittest_component
bringin the test as an executable. The reason is that the shard was
being added in the test(). This should have been done in
fuchsia_unittest_component.

Tested: verified that we see logs in the test we were concerned about
Bug: 319495184
Change-Id: Ie7e63392da90d6bf6412fffbc5ead5a9f437333e
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/972152
Reviewed-by: Ankur Mittal <anmittal@google.com>
Reviewed-by: Gary Bressler <geb@google.com>
Commit-Queue: Miguel Flores <miguelfrde@google.com>
Reviewed-by: Aaron Wood <aaronwood@google.com>