[zbi_test] Introduce `disabled` parameter
Parameter acts as an alias of setting device_types=[] such that the
actual meaning is clear. That is, build this zbi and do not run it as a
boot test.

This is commonly used when the zbi is acting as an artifact to a host
interaction test.

Change-Id: I85b6b38eb944e23fa2d29a938496d305ece25a50
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/970946
Reviewed-by: Roland McGrath <mcgrathr@google.com>
Reviewed-by: Cameron Finucane <eep@google.com>
Commit-Queue: Gianfranco Valentino <gevalentino@google.com>