[kernel_zbi_test] Generate cuckoo targets automatically
Building up on `zbi_test()` forwarding `data_deps` to `test_spec()`
and not to `zbi()`, this changes introduces automatic generation of
cuckoo targets for any `kernel_zbi_test()`.

The default behavior is controlled by the new build argument
`kernel_zbi_test_generates_cuckoo` which is false. This can be overridden
in a per target fashion.

Ideally this flag is only set in coverage bots.

As an example the standalone-coverage-verification-test now uses this
flag instead of manually specifying the cuckoo target.

Bug: 42052700
Change-Id: I360c203ee65120b2051c3acb67963c63d9893fd9
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/967260
Reviewed-by: Joshua Seaton <joshuaseaton@google.com>
Reviewed-by: Aidan Wolter <awolter@google.com>
Reviewed-by: Roland McGrath <mcgrathr@google.com>
Commit-Queue: Gianfranco Valentino <gevalentino@google.com>