[starnix] Use component events to fence coredump test.
All the flakes I looked at were timeouts on slow bots that are often
contended. My hypothesis is that the timeouts are because the test
relied on seeing values show up in Inspect, and it would sleep for a
second if they weren't present. This happened at least once for every
iteration of the test.

Using component events allows us to immediately read the kernel's
inspect without needing to potentially retry after a delay because
Starnix has already recorded the coredump in inspect by the time it
has closed the component controller handle.

I'm not positive that this will address all sources of the flake, but
I'm going to mark the bug as fixed so that any failures after this
lands can be tracked separately.

Multiply: starnix_coredump_diagnostics_test: 100
Fixed: b/317401801
Change-Id: Ia77d1f381142c1354490881fcc5f1536d58b58aa
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/971810
Reviewed-by: Vickie Cheng <vickiecheng@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>
Fuchsia-Auto-Submit: Adam Perry <adamperry@google.com>