[netstack3] Tidy up context markers
Implement CoreContext marker based only on IpBindingsContext as its
dependency, and IpBindingsContext is itself used to define
BindingsContext.

To do that, cleanup places that were using BindingsContexts bounds
and use BindingsTypes instead. Notably TimerContext conversions.

This allows bindings to enrich its impl blocks with bounds like
`where BindingsCtx: IpBindingsContext<I>`, which satisfies the compiler
locally. Otherwise, the I bounds on *all* implementations on BindingsCtx
would have to be the same for the compiler to be able to resolve that it
implements it. This solution is more verbose but less surprising in the
long run.

Change-Id: Icf8857cadeb4e3fc115844c40d24a0522232f60d
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/971882
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>
Reviewed-by: Evan Wildenhain <wildenhain@google.com>
Fuchsia-Auto-Submit: Bruno Dal Bo <brunodalbo@google.com>