[display][dsi] Delete the dsiutil command line tool.
The dsiutil command line tool was created to bypass Scenic
and the display coordinator to control the MIPI-DSI LCD.

This was used for debugging but this made it hard for drivers
and users to reason about the whole display stack as users
can control the LCD's display driver IC arbitrarily.

Given that there's no active user of the dsiutil driver, we
decide to delete the dsiutil binary from our builds.

Test: no-op.
Bug: 319379442
Change-Id: I40ff6bafd52714f13698a5886086f31bfa425dff
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/971096
Fuchsia-Auto-Submit: Yilong Li <liyl@google.com>
Reviewed-by: Victor Costan <costan@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>
Reviewed-by: Aaron Wood <aaronwood@google.com>