[drivers] Fix issues with improper driver shutdown
Add the value of `removal_set` to the in-function stack. This
fixes as issue where it is possible that boot drivers get shut
down too early. This happens because while traversing the device
topology we can set `removal_set` to `kAll`. The `kAll` value
then persists as we move back up the in-function stack, which can
lead to incorrect decisions.

Fixed: b/316925341
Change-Id: I9f8e2e18f825d5d70ba4fceeebc8bd70f2a01fe6
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/968782
Commit-Queue: Justin Mattson <jmatt@google.com>
Reviewed-by: Garratt Gallagher <garratt@google.com>
Reviewed-by: Sarah Chan <spqchan@google.com>